fix: duplicate operation ids from additional_bindings (#92)
The issue was that when a proto method has `additional_bindings`, the resulting OpenAPI specification would have different paths with the same operation ID. This violates the OpenAPI specification which requires operation IDs to be unique.

My solution is to append a number when there are duplicates, starting at 2 and leaving the first instance unchanged so existing operationIds are unaffected.

Closes #88